requirements.txt
annotated-types==0.7.0
bertopic==0.16.2
blis==0.7.11
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
cloudpickle==3.0.0
confection==0.1.5
cuda-python==12.5.0
cudf-cu12==24.6.1
cuml-cu12==24.6.1
cupy-cuda12x==13.2.0
cymem==2.0.8
Cython==0.29.37
dask==2024.5.1
dask-cuda==24.6.0
dask-cudf-cu12==24.6.1
dask-expr==1.1.1
distributed==2024.5.1
distributed-ucxx-cu12==0.38.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
fastrlock==0.8.2
filelock==3.15.4
fsspec==2024.6.1
hdbscan==0.8.37
huggingface-hub==0.23.4
idna==3.7
importlib_metadata==8.0.0
Jinja2==3.1.4
joblib==1.4.2
kaleido==0.2.1
langcodes==3.4.0
language_data==1.2.0
libucx-cu12==1.15.0.post1
llvmlite==0.43.0
locket==1.0.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.8
murmurhash==1.0.10
networkx==3.3
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
nvtx==0.2.10
packaging==24.1
pandas==2.2.2
partd==1.4.2
pillow==10.4.0
plotly==5.22.0
preshed==3.0.9
psutil==6.0.0
pyarrow==16.1.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pylibraft-cu12==24.6.0
pynndescent==0.5.13
pynvjitlink-cu12==0.3.0
pynvml==11.4.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
raft-dask-cu12==24.6.0
rapids-dask-dependency==24.6.0
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rmm-cu12==24.6.0
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
sentence-transformers==3.0.1
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sortedcontainers==2.4.0
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.13.0
tblib==3.0.0
tenacity==8.5.0
thinc==8.2.5
threadpoolctl==3.5.0
tokenizers==0.19.1
toolz==0.12.1
torch==2.3.1
tornado==6.4.1
tqdm==4.66.4
transformers==4.42.4
treelite==4.1.2
triton==2.3.1
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ucx-py-cu12==0.38.0
ucxx-cu12==0.38.0
umap-learn==0.5.6
urllib3==2.2.2
wasabi==1.1.3
weasel==0.4.1
wrapt==1.16.0
zict==3.0.0
zipp==3.19.2
